Remove deprecated driver classes
