--- src/Driver/IBMDB2/DB2Driver.php
@@ -1,27 +1,0 @@
-<?php
-
-namespace Doctrine\DBAL\Driver\IBMDB2;
-
-use Doctrine\DBAL\Driver\AbstractDB2Driver;
-
-/**
- * IBM DB2 Driver.
- *
- * @deprecated Use {@link Driver} instead
- */
-class DB2Driver extends AbstractDB2Driver
-{
-    /**
-     * {@inheritdoc}
-     */
-    public function connect(array $params)
-    {
-        return new Connection(
-            DataSourceName::fromConnectionParameters($params)->toString(),
-            isset($params['persistent']) && $params['persistent'] === true,
-            $params['user'] ?? '',
-            $params['password'] ?? '',
-            $params['driver_options'] ?? []
-        );
-    }
-}
